refoactor: 디렉토리 변경
attendance -> domain
