+++ src/main/java/attendance/view/OutputView.java
@@ -1,4 +1,4 @@
-package view;
+package attendance.view;
 
 public class OutputView {
