+++ src/test/java/attendance/domain/AttendanceHistoryTest.java
@@ -1,4 +1,4 @@
-package attendance;
+package attendance.domain;
 
 
 import static org.assertj.core.api.Assertions.assertThat;
